+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - `Outline`'s inferred direction annotations of `S` joined to over and under
    blends
+- Ranking of search results containing `A&'` or `A&E`
 
 ## 0.9.0 - 2026-04-21